+++ src/match/viterbi.py
@@ -1,3 +1,4 @@
+import imp
 import numpy as np
 import geopandas as gpd
 from loguru import logger
@@ -53,38 +54,48 @@ def viterbi_decode(nodes, trans):
     # 获得输入状态序列的长度，以及观察标签的个数
     seq_len, num_labels = len(nodes), len(trans)
     # 简单起见，先不考虑发射概率，直接用起始0时刻的分数
-    scores = nodes[0].reshape((-1, 1))
+    scores = nodes[0].reshape((-1, 1)) # (num_labels, 1)
     
     paths = []
     # 递推求解上一时刻t-1到当前时刻t的最优
     for t in range(1, seq_len):
         # scores 表示起始0到t-1时刻的每个标签的最优分数
-        scores_repeat = np.repeat(scores, num_labels, axis=1)
+        scores_repeat = np.repeat(scores, num_labels, axis=1) # (num_labels, num_labels)
+        
         # observe当前时刻t的每个标签的观测分数
-        observe = nodes[t].reshape((1, -1))
-        observe_repeat = np.repeat(observe, num_labels, axis=0)
+        observe = nodes[t].reshape((1, -1)) # (1, num_labels)
+        observe_repeat = np.repeat(observe, num_labels, axis=0) # (num_labels, num_labels)
+        
         # 从t-1时刻到t时刻最优分数的计算，这里需要考虑转移分数trans
         M = scores_repeat + trans + observe_repeat
+        
         # 寻找到t时刻的最优路径
         scores = np.max(M, axis=0).reshape((-1, 1))
         idxs = np.argmax(M, axis=0)
+        
         # 路径保存
         paths.append(idxs.tolist())
         
     best_path = [0] * seq_len
     best_path[-1] = np.argmax(scores)
+    
     # 最优路径回溯
     for i in range(seq_len-2, -1, -1):
         idx = best_path[i+1]
         best_path[i] = paths[i][idx]
 
 
+def get_trans_prob(trans_prob, layer_id):
+    return trans_prob[layer_id]
+
+
 def decode(observations, states, start_prob, trans_prob, emit_prob, mode='+'):
-    if mode == '+':
-        f = lambda x, y: x +  y
-    elif mode == '*':
-        f = lambda x, y: x *  y
-        
+    def _formula(x, y):
+        if mode == '+':
+            return x +  y
+        elif mode == '*':
+            return x *  y
+            
     V = [{}]
     path = {}
 
@@ -103,12 +114,14 @@ def decode(observations, states, start_prob, trans_prob, emit_prob, mode='+'):
         for curr_st in states:
             paths_to_curr_st = []
             for prev_st in V[t-1]:
-                if (prev_st, curr_st) not in trans_prob[t-1]:
+                _trans_prob = get_trans_prob(trans_prob, t-1)
+                if (prev_st, curr_st) not in _trans_prob:
                     continue
-                a = V[t-1][prev_st]
-                b = trans_prob[t-1][(prev_st, curr_st)]
-                c = emit_prob[curr_st][observations[t]]
-                paths_to_curr_st.append(( f(a, b * c), prev_st))
+                
+                v = V[t-1][prev_st]
+                _v = _trans_prob[(prev_st, curr_st)]
+                _e = emit_prob[curr_st][observations[t]]
+                paths_to_curr_st.append(( _formula(v, _v * _e), prev_st))
             
             if not paths_to_curr_st:
                 continue
@@ -125,26 +138,44 @@ def decode(observations, states, start_prob, trans_prob, emit_prob, mode='+'):
     return prob, path[end_state]
 
 
-def prepare_viterbi_input(cands, graph):
+def prepare_viterbi_input(cands, gt):
     states = cands.eid.unique()
     observations = cands.pid.unique()
     start_prob = cands.query("pid == 0").set_index('eid')['observ_prob'].to_dict()
     # start_prob = {key:1 for key in start_prob}
-    
-    # BUG cands 坐标不连续的问题, 莫非是中断
-    trans_prob = [graph.loc[i]['f'].to_dict() for i in observations[:-1] ]
 
     observ_dict = cands[['pid', 'eid', 'observ_prob']].set_index(['eid'])
     emit_prob = {i: observ_dict.loc[[i]].set_index('pid')['observ_prob'].to_dict() for i in states}
+
+    # BUG cands 坐标不连续的问题, 莫非是中断
+    trans_prob = [gt.loc[i]['f'].to_dict() for i in observations[:-1] ]
+    
+    # tmp = gt[['d_euc', 'first_step_len', 'last_step_len', 'cost', 'path', 'geometry', 'v', 'first_step', 'last_step', 'move_dir', 'f_dir']]
     
     return states, observations, start_prob, trans_prob, emit_prob
         
 
-def process_viterbi_pipeline(cands, graph):
-    states, observations, start_prob, trans_prob, emit_prob = prepare_viterbi_input(cands, graph)
+def process_viterbi_pipeline(cands, gt):
+    states, observations, start_prob, trans_prob, emit_prob = prepare_viterbi_input(cands, gt)
     _, rList = decode(observations, states, start_prob, trans_prob, emit_prob)
 
-    # BUG
     rList = cands.set_index(['pid', 'eid']).loc[rList][[ 'src', 'dst']].reset_index()
     
     return rList
+
+
+if __name__ == "__main__":
+    import sys
+    sys.path.append('../')
+    from utils.serialization import load_checkpoint
+    
+    fn = "../debug/traj_0_data_for_viterbi.pkl"
+    fn = "../../debug/traj_1_data_for_viterbi.pkl"
+    # fn = Path(__file__).parent / fn
+    data = load_checkpoint(fn)
+
+    cands = data['cands']
+    gt = data['graph']
+    rList = data['rList']
+
+    res = process_viterbi_pipeline(cands, gt)